requirements.txt
Django
whitenoise
gunicorn
djangorestframework